backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
httpx==0.24.0
pydantic==1.10.8
python-multipart==0.0.6
pytest==7.3.1
requests==2.31.0
